% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg.R
\name{addRtoolsPath}
\alias{addRtoolsPath}
\title{Add Rtools to System PATH}
\usage{
addRtoolsPath(rtools.root = NULL)
}
\arguments{
\item{rtools.root}{Where Rtools is installed.}
}
\value{
Nothing
}
\description{
Rtools is very important. If it is not imported to PATH, you need to
run this function to apply. You can manually download and install Rtools and
export them to System PATH.
}
\examples{
\dontrun{
addRtoolsPath()
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
