% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{aseskit}
\alias{aseskit}
\alias{aseskit-package}
\title{aseskit: An R General Analytics Toolkit for ASES}
\description{
An analytic toolkit comprising of a series of generic work functions,
e.g., partial evaluation, distribution analysis, package load/unload, GUI,
iif, ...
}
\details{
This package is comprised of \cr \describe{
  \item{API helpers}{\code{\link{get_api_data}()}, \code{\link{parse_api_data}()},
    \code{\link{synthesize_api}()}, \code{\link{get_api_key}()},
    \code{\link{set_api_key}()}}
  \item{Password}{\code{\link{input_pwd}()}, \code{\link{get_raw_pwd}()},
    \code{\link{decrypt_it}()}, \code{\link{encrypt_it}()}}
  \item{Package mgmt}{\code{\link{load_pkgs}()}, \code{\link{unload_pkgs}()},
    \code{\link{dep_pkgs}()}, \code{\link{addRtoolsPath}()}}
  \item{Vectorized \code{\link{ifelse}}}{\code{\link{iif}()} family}
  \item{Format convertion}{\code{\link{to_pct}()}, \code{\link{to_num}()}}
  \item{Distribution analysis}{\code{\link{describe_df}()},
    \code{\link{cal_df_distrib}()}, \code{\link{vis_df_distrib}()}}
  \item{Helpers}{\itemize{
    \item nchar: \code{\link{count_char}()}
    \item Relabel number range: \code{\link{relab_range}()}
    \item Clock-wise position: \code{\link{clock_to_pos}()}
    \item Parse clipboard: \code{\link{parse_clipb}()}
    \item JS-style operators: \code{\link{\%+=\%}()}
    \item Align vectors: \code{\link{aline}()}
    \item Enquo arguments: \code{\link{enquo_arg}()}
    \item Parse and extract formula: \code{\link{parse_formula}()}
  }}
}
}
\author{
\strong{Creator, Maintainer}: Yiying Wang, \email{wangy@aetna.com}
}
\keyword{internal}
