% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{check_curl}
\alias{check_curl}
\title{Check if curl is avaiable for given API URL(s)}
\usage{
check_curl(url, app_name = gsub("^[^/]+//(.*)$", "\\\\1", url),
  use_curl = TRUE, messaging = FALSE)
}
\arguments{
\item{url}{character vector of API URLs}

\item{app_name}{character vector of API app name. If missing, \code{url} will be
used as the name(s)}

\item{use_curl}{logical, whether to use curl. The function will check the url(s)
only when use_curl = TRUE. Default TRUE.}

\item{messaging}{logical, whether print messages while processing. Default FALSE.}
}
\value{
a logical vector corresponding to url.
}
\description{
Check if curl is avaiable for given API URL(s)
}
\examples{
\dontrun{
check_curl(c("http://www.baidu.com",'https://www.google.com'))  # returns

# www.baidu.com www.google.com
#         FALSE           TRUE
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
