% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/num_pct.R
\name{convNum2Pct}
\alias{convNum2Pct}
\alias{to_pct}
\title{Convert numbers to percents}
\usage{
convNum2Pct(x, digits = 0, value_asis = TRUE, ...)

to_pct(x, digits = 0, value_asis = TRUE, ...)
}
\arguments{
\item{x}{A numeric vector, list, or matrix.}

\item{digits}{Digits of the percentage numbers. Default 0.}

\item{value_asis}{Logical, default TRUE. Whether keep the actual value. If FALSE,
the result will keep the numeric part as is (100 times smaller, e.g., 0.5 --> 0.5\%).}
}
\value{
Character in percentage format, as with the same structure with \code{x}.
}
\description{
Convert a vector, list or matrix of numeric values to percentage values.
}
\examples{
\dontrun{
## Multiply the numbers with 100 and *100\%
to_pct(c(0.04, 0.3, NA), digits=1)  # return
# c("4.0\%", "30.0\%", NA)

to_Pct(c(0.04, 0.3, NA), digits=1, value_asis=FALSE)  # returns
# c("0.04\%", "0.3\%", NA)

to_pct(list(0.04, -0.9, list(0.65, NULL)))
[[1]]
[1] "4\%"

[[2]]
[1] "-90\%"

[[3]]
[[3]][[1]]
[1] "65\%"
[[3]][[2]]
NULL
}
}
\seealso{
\code{\link[scales]{number_format}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
