% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nchar.R
\name{countChar}
\alias{countChar}
\alias{count_char}
\title{Get Number of Characters}
\usage{
countChar(x, rm.blank = TRUE, ...)

count_char(x, rm.blank = TRUE, ...)
}
\arguments{
\item{x}{Object. Vector, matrix, list or data frame.}

\item{rm.blank}{Logical: default TRUE, and '' will be coerced to NA.}

\item{...}{Other arguments to pass to \code{.countChar}}
}
\value{
An object of the same structure with \code{x}.
}
\description{
Get the number of characters in an object. It supports recursive counting for 
lists.
}
\examples{
\dontrun{
count_char(c("dws", 23)) # return c(3, 2)

count_char(matrix(c("dws", 23, "2332", 1), ncol=2))  # return
##       [,1] [,2]
## [1,]    3    4
## [2,]    2    1

count_char(list("dws", 23, c("2332", 1)))  # return
## [[1]]
## [1] 3
##
## [[2]]
## [1] 2
##
## [[3]]
## [1] 4 1
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
