% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data:CLOCKPOS}
\alias{data:CLOCKPOS}
\title{Dataset: CLOCK POS}
\description{
Clock position vectors (horizontal, vertical and direction) of the \code{clock_to_pos}
function.
}
\details{
It is a 26 * 3 character matrix: \itemize{
\item horizontal: "right", "left" or "center"
\item vertical: "top", "bottom" or "center"
\item direction: "horizontal" or "vertical"
}
}
\examples{
\dontrun{
data(CLOCKPOS)
}
}
\references{
The data set is used for \code{\link{clock_to_pos}}
}
\author{
Yiying Wang \email{wangy@aetna.com}
}
\keyword{clock}
\keyword{data}
\keyword{position}
