% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dep_pkg.r
\name{dep_pkgs}
\alias{dep_pkgs}
\alias{dep_pkg}
\title{Get dependent packages for the given packages}
\usage{
dep_pkgs(..., dep_type = c("Depends", "Imports", "LinkingTo"),
  repos = getOption("repos"), recursive = FALSE)

dep_pkg(..., dep_type = c("Depends", "Imports", "LinkingTo"),
  repos = getOption("repos"), recursive = FALSE)
}
\arguments{
\item{...}{package names for process. Can be in the following forms: \describe{
\item{character vector/list}{E.g., c("tidyr", "dplyr") or list("tidyr", "dplyr")}
\item{characters}{E.g., "tidyr", "dplyr"}
\item{symbols}{E.g., tidyr, dplyr}
}}

\item{dep_type}{character, type of dependency, accecpts c("Depends", "Imports",
"LinkingTo", "Suggests", "Enhances"). Default c('Depends', 'Imports', 'LinkingTo').
Character string "all" is shorthand for that vector, character string "most"
for the same vector without "Enhances".}

\item{repos}{repos address, default \code{getOption("repos")}}

\item{recursive}{logical, whether recursive when fetching package dependencies.
Default FALSE.}
}
\value{
a tibble of dependent/import packages of the given \code{packages}. It has
4 columns: \describe{
\item{package}{dependent package names}
\item{cran}{package version on CRAN}
\item{local}{package version locally}
\item{behind}{whether local version is behind CRAN}
}
}
\description{
Get a tibble of packages that the given packages depend on or import.
It makes use of \pkg{utils}::\code{\link[utils]{available.packages}} to get the full
package meta info list.
}
\details{
This function requires internet connection.
}
\examples{
\dontrun{
dep_pkgs(c("dplyr", "devtools"))  # or
dep_pkgs("dplyr", "devtools")
}
}
\seealso{
\code{\link[tools]{package_dependencies}}
}
