% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distrb.R
\name{describe_df}
\alias{describe_df}
\alias{cal_df_distrib}
\alias{vis_df_distrib}
\title{Describe and visulaize the statistic distribution of a data.frame}
\usage{
describe_df(data, qual_vars, quan_vars, nrow = 1, ncol = 1, ...)

cal_df_distrib(data, qual_vars, quan_vars, ...)

vis_df_distrib(data, qual_vars, quan_vars, nrow = NULL, ncol = NULL,
  ...)
}
\arguments{
\item{data}{a data.frame}

\item{qual_vars}{variable(s) indicating the qualitative variable(s).
It accepts five forms: \itemize{
\item missing (default), the function will automatically identify the qualitative variable
\item character vector indicating variable names, e.g., \code{c("cyl", "gear")}
\item integer vector indicating variable index, e.g., \code{c(2, 10)}
\item quosures yielded using \code{\link[dplyr]{vars}}(), e.g, \code{vars(cyl, gear)}
\item NULL indicating that the function will not process qualitatitive variables
}}

\item{quan_vars}{variable(s) indicating the quantitative variable(s).
It accepts five forms: \itemize{
\item missing (default), the function will automatically identify the quantitative variables
\item character vector indicating variable names, e.g., \code{c("drat", "mpg")}
\item integer vector indicating variable index, e.g., \code{c(1, 5)}
\item quosures yielded using \code{\link[dplyr]{vars}}(), e.g, \code{vars(drat, mpg)}
\item NULL indicating that the function will not process quantitative variables
}}

\item{nrow}{NULL or positive integer. Only applicable for \code{vis_df_distrib},
determining the number of rows in the plot grid. See \code{\link[ggplot2]{facet_wrap}}
for more details.}

\item{ncol}{NULL or positive integer. Only applicable for \code{vis_df_distrib},
determining the number of columns in the plot grid. See \code{\link[ggplot2]{facet_wrap}}
for more details.}

\item{...}{other arguments to pass to \code{describe_df}, \code{cal_df_distrib}
and \code{vis_df_distrib}. Accepts the following:
\itemize{
 \item pass to \code{cal_df_distrib} \describe{
  \item{argument for \code{choose_col_idx}}{diversity_threshold}
  }
 \item pass to \code{vis_df_distrib} \describe{
  \item{\code{help_on_dots}}{logical. If TRUE, you can get help info for
    \code{...} argument.}
  \item{elipsis arguments for \code{\link[ggplot2]{geom_histogram}}}{hist.title,
    hist.tag, hist.subtitle, hist.stat, hist.position, hist.bins, hist.color,
    hist.inherit.aes}
  \item{elipsis arguments for histograms and \code{\link[ggplot2]{geom_bar}}}{
    \code{bar.stat, bar.position, bar.width, bar.binwidth, bar.fill, bar.color},
    \code{bar.size, bar.alpha, bar.na.rm, bar.show.legend, bar.inheirt.aes}}
  \item{elipsis arguments for \code{\link[ggplot2]{geom_vline}}}{
    \code{vline.color, vline.size, vline.alpha, vline.show.legend, vline.na.rm}}
  }
}}
}
\value{
\code{describe_df}(): a list of results by \code{cal_df_distrib} and
\code{vis_df_distrib}

\code{cal_df_distrib}(): two lists of tibbles \describe{
  \item{'qual'}{a tibble with 4 columns: \code{<varname>, 'value', 'freq', 'prop'};
  or a character \code{"no match columns"}}
 \item{'quan'}{a tibble with 11 columns: \code{<varname>, 'count', 'n_na', 'p_na'},
   \code{'mean', 'sd', 'min', 'lower', 'median', 'higher', 'max'}; or a character
   \code{"no match columns"}}
}

\code{vis_df_distrib}(): two sets of plots \itemize{
 \item \code{nrow} and \code{ncol} are 1 respectively, then the distribution
  plots (histogram or bar plots) will be shown one by one. \cr
 \item either \code{nrow} or \code{ncol} is more than 1, then the plots will
  be organized in facetted plots (with \code{\link[ggplot2]{facet_wrap}()}). \cr
 \item If both \code{nrow} and \code{ncol} are NULL, then the plots will be
  organized into one plot
}
}
\description{
A family of functions to analyze a data.frame and output the
descriptive analysis results and/or visualization.
}
\details{
\code{describe_df} is a wrapper of \code{cal_df_distrib} and
\code{vis_df_distrib}. \itemize{
 \item \code{cal_df_distrib}, built on top of \code{describe_num_cols} and
  \code{describe_chr_cols}, returns two lists: one for quantitative variables
  and the other for qualitative ones. \cr
 \item \code{vis_df_distrib}, built on top of \code{vis_num_cols} and
  \code{vis_chr_cols}, returns two sets of plots: one for quantitative variables
  and the other for qualitative ones, each of which contains one or multiple
  \pkg{ggplots} objects.
}
}
\examples{
\dontrun{
describe_df(mtcars, vars(am, cyl, mpg, wt), NULL)

# you can also detect the distribution of the data in clipboard
describe_df(parse_clipb())
}
\dontrun{
cal_df_distrib(iris)  ## or
library(dplyr)
cal_df_distrib(iris, quan_vars=vars(Sepal.Length, Petal.Length))
}
\dontrun{
library(dplyr)
vis_df_distrib(mtcars, qual=vars(am, cyl), quan=vars(mpg, wt))
}
}
