% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwd.R
\name{encryptIt}
\alias{encryptIt}
\alias{decryptIt}
\alias{decrypt_it}
\title{Encrypt a character and decrypt a raw vector}
\usage{
encryptIt(x, ...)

decryptIt(x, ...)

decrypt_it(x, ...)
}
\arguments{
\item{x}{the raw vector to decrypt}

\item{...}{other arguments.}

\item{mode}{character, 'ECB" or 'CBC'. Default 'CBC', which is more secure.}

\item{IV}{NULL or a numeric vector, the initial vector for 'CBC' or initial
counter for 'CTR' mode in \pkg{digest}::\code{\link[digest]{AES}()}. It is
bypassed when \code{mode} == 'ECB'. Default NULL.}
}
\value{
invisible
}
\description{
It calls a builtin algorithm to encrypt a character or decrypt a raw vector.
If \code{x} is not of the \code{encrypt} class, the value will not be decrypted.
}
\examples{
\dontrun{
encrypt_it('my key')
}
\dontrun{
decryptIt("Pwd")
}
}
