% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwd.R
\name{getRawPwd}
\alias{getRawPwd}
\alias{get_raw_pwd}
\title{Get ACC password (in raw) from the acckey.pem file}
\usage{
getRawPwd(pem = "acckey", overide = FALSE)

get_raw_pwd(pem = "acckey", overide = FALSE)
}
\arguments{
\item{pem}{character, the name of the .pem file. Default 'acckey'.}

\item{overide}{logical, if overide existing acckey.pem and input password again.
Default FALSE.}
}
\value{
invisible (an encrypted raw vector)
}
\description{
If the acckey.pem file already exists, it extracts the
cipher text from it and decrypt it into key string. Otherwise, it will call
\code{inputPwd} to enter the password. \cr
Note that this function returns \strong{an encrypted raw vector} which needs
to be decrypted using \code{\link{decrypt_it}()}.
}
\examples{
\dontrun{
qry_acc <- function(dbname, query) {
    con <- DBI::dbConnect(odbc::odbc(), .connection_string=paste0(
        "Driver={Microsfot Access Driver (*.mdb, *.accdb)};",
        "Dbq=", dbname, ";uid=;pwd=", decrypt_it(get_raw_pwd())))
    return(DBI::dbGetQuery(con, query))
}
qry_acc("testdb.accdb", query="select * from table1")
}

}
\seealso{
\code{\link{inputPwd}}  \code{\link{decrypt_it}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
