% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui_wizard.R
\name{inputChar}
\alias{inputChar}
\alias{input_char}
\title{Input a character}
\usage{
inputChar(prompt = NULL, caption = NULL, default = "",
  gui = c("Rstudioapi", "Gwidgets", "WinGui"), ...)

input_char(prompt = NULL, caption = NULL, default = "",
  gui = c("Rstudioapi", "Gwidgets", "WinGui"), ...)
}
\arguments{
\item{prompt}{character, the prompt label of the GUI window. Default NULL.}

\item{caption}{character, the caption of the GUI window. Default NULL.}

\item{default}{character, the default value for the inputbox. Default "".}

\item{gui}{'Rstudioapi', 'Gwidgets' or 'WinGui'. Default 'Rstudioapi'. If 
'Rstudioapi' and Rstudio version >= 1.1.67, then call 
\code{\link[rstudioapi]{showPrompt}}, or it calls \code{\link{winDialogString}} 
in Windows OS, otherwise uses a GUI wizard with gWidget2,}

\item{...}{other arguments.}
}
\value{
invisible
}
\description{
Input a character via GUI wizard. It returns nothing to the screen unless you 
assign the output to a variable.
}
\examples{
\dontrun{
x <- input_char(prompt="Input the api key", caption="API Key?")
}

}
\seealso{
\code{\link[rstudioapi]{showPrompt}}  \code{\link[utils]{winDialogString}}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
