% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pwd.R
\name{inputPwd}
\alias{inputPwd}
\alias{input_pwd}
\title{Input password key}
\usage{
inputPwd(prompt = NULL, caption = NULL, encrypt = TRUE,
  gui = c("Rstudioapi", "Gwidgets"), gen_pem = TRUE, ...)

input_pwd(prompt = NULL, caption = NULL, encrypt = TRUE,
  gui = c("Rstudioapi", "Gwidgets"), gen_pem = TRUE, ...)
}
\arguments{
\item{prompt}{character, the prompt label of the GUI window.}

\item{caption}{character, the caption of the GUI window.}

\item{encrypt}{logical, if the password will be encrypted (a raw vector).
Default TRUE.}

\item{gui}{'Rstudioapi', or 'Gwidgets', default 'Rstudioapi'. If
'Rstudioapi' and Rstudio version >= 1.1.419, then call
\code{\link[rstudioapi]{askForPassword}}, otherwise use a GUI wizard with gWidget2.}

\item{gen_pem}{logical, indicating whether generate a acckey.pem file to store
the cypher key of the password in under \code{\link{tempdir}()}. Default TRUE.}

\item{...}{other arguments, \itemize{
 \item force: set \code{force} = TRUE to overwrite the .pem file forcefully \cr
 \item pem: assign a name to generate the .pem file in \code{dir}. Default
   'acckey'.
}}
}
\value{
invisible
}
\description{
Input a password phrase. It returns nothing unless you assign the output to a
variable.
}
\examples{
\dontrun{
input_pwd(prompt="Input the db key", caption="Password")
}

}
\seealso{
\code{\link[rstudioapi]{askForPassword}}  \code{\link[rstudioapi]{askForSecret}}
You can also use \code{\link{getPwd}} directly
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
