% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabel_range.R
\name{renameRange}
\alias{renameRange}
\alias{relab_range}
\title{Rename data range labels}
\usage{
renameRange(x, quick.form = c("-", "~", "[)", "(]", "<=", ">="),
  format = NULL, left.subtract = 0, right.subtract = 0, ...)

relab_range(x, quick.form = c("-", "~", "[)", "(]", "<=", ">="),
  format = NULL, left.subtract = 0, right.subtract = 0, ...)
}
\arguments{
\item{x}{The factor vector yielded by \code{\link{cut}} function.}

\item{quick.form}{A string to define the new format.
\describe{
 \item{-}{equivalent to \code{format="x-y"}: 0-10, 10-20, 20-30}
 \item{~}{equivalent to \code{foramt="x~y"}: 0~10, 10~20, 20~30}
 \item{[)}{equivalent to \code{foramt="[x, y)"}: [0, 0), [10, 0), [20, 0)}
 \item{(]}{equivalent to \code{foramt="(x, y]"}: (0, 0], (10, 0], (20, 0]}
 \item{<=}{equivalent to \code{foramt=c("<=y", "x-y", ">x")}: <=9, 10-19, 20-29, >29}
 \item{>=}{equivalent to \code{foramt=c("<y", "x-y", ">=x")}: <10, 10-20, 20-30, >=30}
}}

\item{format}{Self-defined character vector length 1-3 to format the label. If 
more than 3 character elements are given, the function will only use the first 3.
Default NULL, and \code{quick.form} is applied. See details.}

\item{left.subtract}{Integer which is subtracted from the left limit of
each range. Default 0.}

\item{right.subtract}{Integer which is subtracted from the right limit of
each range. Default 0.}

\item{...}{ignore}
}
\value{
A new factor vector with reformatted levels.
}
\description{
We usually use \code{\link{cut}} to cut continuous vectors into factors. But the levels
are typically presented as \code{'[l, u)'} or \code{'(l, u]'}. This function can reformat
the cut results.
}
\details{
All the \code{format} definitions should use "x" to represent the lower
limit and "y" to reresent the upper limit of each range label. For instance, 
'x-y' coereces '(0, 1]' to '0-1', '<y' coerces '(0, 1]' to '<1'. \cr 

By principle, it should be of length 3: beginning label, middle label(s) and 
end label. But there could be shortened forms: \itemize{
 \item length 1: all of the same format. E.g., \code{'x-y'} results in '0-1',
 '1-2', '2-3'. \cr
 \item length 2: beginning label and middle label(s). End label is the same with
 middle label. E.g, \code{c("<x", "x-y") results in "<0", "1-2", "2-3".} \cr
 \item length 3: full-form \code{format}.
}
}
\examples{
\dontrun{
x = cut(1:100, c(0, 20, 40, 60, 80, 100))
renameRange(levels(x), "-")
# [1] "0-20"   "20-40"  "40-60"  "60-80"  "80-100"

renameRange(levels(x), "[)")
# [1] "[0, 20)"   "[20, 40)"  "[40, 60)"  "[60, 80)"  "[80, 100)"

renameRange(levels(x), ">=", left.substract = -1)
# [1] "<20"   "21-40"  "41-60"   "61-80"  "\\u226581"
 
relab_range(levels(x), format=c(" ~ y", "x ~ y", "x ~ "))
# [1] " ~ 20"   "20 ~ 40" "40 ~ 60" "60 ~ 80" "80 ~ "  
}
}
\author{
Yiying Wang, \email{wangy@aetna.com}
}
