% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{synthesize_apis}
\alias{synthesize_apis}
\alias{synthesize_api}
\title{Synthesize API URLs}
\usage{
synthesize_apis(url_body, provider = c("google", "baidu", "gaode",
  "ipstack", "ipinfo", "ipify"), api = c("geocode", "revgeocode",
  "convcoord", "geohost"), ...)

synthesize_api(url_body, provider = c("google", "baidu", "gaode",
  "ipstack", "ipinfo", "ipify"), api = c("geocode", "revgeocode",
  "convcoord", "geohost"), ...)
}
\arguments{
\item{url_body}{character, the main body of the caller API.}

\item{provider}{character, 'google', 'baidu', 'gaode', 'ipstack', 'ipinfo', or 'ipify'}

\item{api}{character, 'geocode', 'revgeocode', 'convcoord', or 'geohost'}

\item{...}{other arguments to pass to the function}
}
\value{
A list of API URLs, each of which is a \code{url} object.
}
\description{
A wrapper function to synthesize API URL(s) for extraction of data. You can then
use \code{\link{get_api_data}()} to extract the data.
}
\examples{
\dontrun{
synthesize_apis('https://maps.googleapis.com/maps/api/', 'google', 'geocode')
}

}
